% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_id.R
\name{get_game_ids}
\alias{get_game_ids}
\title{Get Game Ids for MLB}
\usage{
get_game_ids(date_start, date_end = 0)
}
\arguments{
\item{date_start}{The starting date}

\item{date_end}{The ending date. Optional, if you only want to scrape for a single day}
}
\value{
game_ids between starting date and ending date
}
\description{
Get Game Ids for MLB
}
\examples{
get_game_ids(date_start = "2020-7-23")
get_game_ids(date_start = "2020-7-23", date_end = "2020-7-24")
}
