% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pitchfx.R
\name{get_pitchfx}
\alias{get_pitchfx}
\title{Scrape PITCHf/x data}
\usage{
get_pitchfx(date_start = 0, date_end = 0, game_id = NULL)
}
\arguments{
\item{date_start}{The starting date. Optional, if you only want to scrape specific games}

\item{date_end}{The ending date. Optional, if you only want to scrape for a single day}

\item{game_id}{Game ids. Optional, if you want to scrape given days}
}
\value{
PITCHf/x data for within starting date and ending date
}
\description{
Scrape PITCHf/x data
}
\examples{
get_pitchfx(date_start = "2020-7-23")
get_pitchfx(date_start = "2020-7-23", date_end = "2020-7-24")
get_pitchfx(game_id = "630851")
}
