\name{getFieldNames}
\alias{getFieldNames}
\title{Find the most complete set of "tags" for an XML node in a given set of urls.}
\usage{
  getFieldNames(urls, node = "game")
}
\arguments{
  \item{urls}{}

  \item{node}{name of XML node}
}
\description{
  The field names should be extracted before the data is
  scraped
}

