\name{getMiniScoreboards}
\alias{getMiniScoreboards}
\title{Generate miniscoreboard urls.}
\usage{
  getMiniScoreboards(first.date, last.date)
}
\arguments{
  \item{first.date}{date (yyyy-mm-dd) to commence scraping
  of pitch F/X data}

  \item{last.date}{date (yyyy-mm-dd) to terminate scraping
  pitch F/X data}
}
\description{
  Generate miniscoreboard urls.
}

