\name{getPitchFxURLs}
\alias{getPitchFxURLs}
\title{Generate Pitch F/X URLs.}
\usage{
  getPitchFxURLs(urls)
}
\arguments{
  \item{urls}{"miniscoreboard" urls used to construct the
  relevant URLs for each game.}
}
\value{
  Returns a list of URLs.
}
\description{
  This function will create the relevant URLs for each
  game.
}

