\name{nodesToDataFrame}
\alias{nodesToDataFrame}
\title{Nodes to Data Frame}
\usage{
  nodesToDataFrame(nodes, tags)
}
\arguments{
  \item{nodes}{XML nodes to manipulate}

  \item{tags}{field names for a specific table}
}
\description{
  This function takes on a set of XML nodes and returns a
  data frame.
}
\details{
  Used to avoid redundancy within the urlsToDataFrame
  function.
}

