\name{animateFX}
\alias{animateFX}
\title{Animate Pitch F/X}
\usage{
  animateFX(data, color, facets, time.interval = 0.01)
}
\arguments{
  \item{data}{pitch F/X data to be visualized.}

  \item{color}{variables within the data set that determine
  different colors}

  \item{facets}{variables within the data set that
  determine different facets}

  \item{time.interval}{time interval between flushing of
  plot}
}
\value{
  ggplot2 object
}
\description{
  Pitch trajectories animated on a two-dimensional plot.
}
\details{
  Details to go here.
}

