\name{urlsToDataFrame}
\alias{urlsToDataFrame}
\title{Scraping work horse}
\usage{
  urlsToDataFrame(urls, p.tags, ab.tags)

  urlsToDataFrame(urls, p.tags, ab.tags)
}
\arguments{
  \item{urls}{URLs for parsing}

  \item{p.tags}{field names for the pitch F/X table}

  \item{ab.tags}{field names for the atbats table}
}
\description{
  Scraping work horse

  This function takes on a list of URLs and returns a data
  frame
}

