\name{urlsToDataFrame}
\alias{urlsToDataFrame}
\title{Scraping Work Horse}
\usage{
  urlsToDataFrame(urls, p.tags, ab.tags)
}
\arguments{
  \item{urls}{vector of strings describing web addresses}

  \item{p.tags}{something}

  \item{ab.tags}{something else export}
}
\description{
  this function goes out to the website and pulls data
}
\details{
  more elaborate description here
}

